asignacion_api/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-filter
django_extensions
requests
Pillow
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
psycopg2-binary
dj-database-url==2.1.0
mssql-django
pyodbc
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
whitenoise==6.7.0
Faker==25.2.0
python-decouple==3.8